+++ docs/changes.rst
@@ -28,6 +28,7 @@ Unreleased.
  - Added :meth:`Image.blue_shift() <wand.image.BaseImage.blue_shift>` method.
  - Added :meth:`Image.color_matrix() <wand.image.BaseImage.color_matrix>` method.
  - Added :meth:`Image.colorize() <wand.image.BaseImage.colorize>` method.
+ - Added :attr:`Image.fuzz <wand.image.BaseImage.fuzz>` property.
  - Added :meth:`Image.remap() <wand.image.BaseImage.remap>` method.
  - Added :meth:`Image.sparse_color() <wand.image.BaseImage.sparse_color>` method.